+++ tests/conftest.py
@@ -21,25 +21,25 @@ conn_params = ConnParams().get()
 commuter = Commuter(**conn_params)
 
 
-def delete_table(table_name, schema='public'):
-    if commuter.is_table_exist(table_name, schema=schema):
+def delete_table(table_name):
+    if commuter.is_table_exist(table_name):
         commuter.execute(
-            sql.SQL("DROP TABLE {} CASCADE").format(
-                sql.Identifier(schema, table_name)))
+            sql.SQL("DROP TABLE IF EXISTS {} CASCADE").format(
+                sql.SQL(table_name)))
 
 
-def with_table(table_name, create_callback, *create_args, schema='public'):
+def with_table(table_name, create_callback, *create_args):
     def decorator(func):
         @wraps(func)
         def wrapped(*args, **kwargs):
-            delete_table(table_name, schema=schema)
+            delete_table(table_name)
             try:
-                cmd = create_callback(table_name, *create_args, schema=schema)
+                cmd = create_callback(table_name, *create_args)
                 commuter.execute(cmd)
                 func(*args, **kwargs)
             except Exception:
                 assert False
             finally:
-                delete_table(table_name, schema=schema)
+                delete_table(table_name)
         return wrapped
     return decorator