+++ tests/test_listener.py
@@ -6,24 +6,23 @@ from .conftest import conn_params, commuter, with_table
 listener = Listener(**conn_params)
 
 
-def create_table(table_name, schema='public'):
-    return f"CREATE TABLE {schema}.{table_name} (id integer, name text)"
+def create_table(table_name):
+    return f"CREATE TABLE {table_name} (id integer, name text)"
 
 
-@with_table('people', create_table, schema='model')
-@with_table('test', create_table, schema='model')
+@with_table("model.people", create_table)
+@with_table("model.test", create_table)
 def test_poll():
     listener.create_notify_function(
-        func_name='notify_trigger',
-        channel='test_channel',
-        schema='model')
+        func_name="model.notify_trigger",
+        channel="test_channel")
 
     listener.create_trigger(
-        table_name='model.people',
-        func_name='notify_trigger')
+        table_name="model.people",
+        func_name="notify_trigger")
 
     listener.poll(
-        channel='test_channel',
+        channel="test_channel",
         on_notify=on_notify,
         on_timeout=on_timeout,
         on_close=on_close,
@@ -31,36 +30,33 @@ def test_poll():
 
     df = commuter.select("SELECT * FROM model.test")
 
-    assert df['id'].to_list() == [2, 3]
+    assert df["id"].to_list() == [2, 3]
 
 
 def on_notify(payload):
     if len(payload) > 0:
         payload = json.loads(payload)
-        _id, _name = int(payload['id']), payload['name']
+        _id, _name = int(payload["id"]), payload["name"]
     else:
-        _id, _name = 2, 'Yeltsin'
+        _id, _name = 2, "Yeltsin"
 
     commuter.insert_row(
-        table_name='test',
+        table_name="model.test",
         id=_id,
-        name=_name,
-        schema='model')
+        name=_name)
 
     raise KeyboardInterrupt
 
 
 def on_timeout():
     commuter.insert_row(
-        table_name='people',
+        table_name="model.people",
         id=1,
-        name='Yeltsin',
-        schema='model')
+        name="Yeltsin")
 
 
 def on_close():
     commuter.insert_row(
-        table_name='test',
+        table_name="model.test",
         id=3,
-        name='Yeltsin',
-        schema='model')
+        name="Yeltsin")